python_ai/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
httpx==0.28.1
PyJWT==2.10.1

# LangChain — agentic loop with tool-calling
langchain==0.3.27
langchain-core==0.3.78
langchain-openai==0.3.35       # Gemini via OpenAI-compatible endpoint
langchain-google-genai==2.1.10 # kept for embeddings only

# Embeddings come from langchain-google-genai (Gemini gemini-embedding-001).
# Vectors are persisted in Chroma Cloud — per-tenant collection, real RAG
# across requests. Local numpy fallback kept for offline dev.
numpy>=1.26,<3.0
chromadb==1.5.8
